requirements.txt
Flask==3.0.2
Werkzeug==3.0.1
qrcode==7.4.2
Pillow==11.2.1
plyer==2.1.0
customtkinter==5.2.2
darkdetect==0.8.0
cryptography==42.0.5
Flask-Session==0.7.0
psutil==5.9.8
Flask-Limiter==3.5.1
schedule==1.2.1
cachelib==0.13.0 